Refactor data_type into directory
Also remove fallback dtypes
