+++ src/data_type/float.rs
@@ -1,0 +1,62 @@
+use std::fmt::Display;
+
+use serde::{Deserialize, Serialize};
+use serde_with::serde_as;
+
+use super::NBytes;
+
+#[serde_as]
+#[derive(Serialize, Deserialize, Debug, Clone, Copy, PartialEq, Eq)]
+#[allow(non_camel_case_types)]
+pub enum FloatSize {
+    // b16,
+    b32,
+    b64,
+}
+
+impl Display for FloatSize {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", self.nbits())
+    }
+}
+
+impl TryFrom<usize> for FloatSize {
+    type Error = &'static str;
+
+    fn try_from(value: usize) -> Result<Self, Self::Error> {
+        match value {
+            // 16 => Ok(Self::b16),
+            32 => Ok(Self::b32),
+            64 => Ok(Self::b64),
+            _ => Err("not a valid float size"),
+        }
+    }
+}
+
+impl NBytes for FloatSize {
+    fn nbytes(&self) -> usize {
+        match self {
+            // Self::b16 => 2,
+            Self::b32 => 4,
+            Self::b64 => 8,
+        }
+    }
+}
+
+// enum SpecialF32 {
+//     NaN,
+//     Infinity,
+//     NegInfinity,
+//     Value(f32),
+//     Bytes([u8; 4]),
+// }
+
+// impl Into
+
+// enum SpecialF64 {
+//     NaN,
+//     Infinity,
+//     NegInfinity,
+//     Value(f64),
+//     Bytes([u8; 8]),
+// }
